requirements.txt
python-dateutil==2.2
six==1.6.1
wsgiref==0.1.2
ystockquote==0.2.5dev
